# Improve performance of System.Random.LegacyImpl.NextBytes(Span<T>)